api/health.py
import logging
from flask import jsonify
from werkzeug.exceptions import InternalServerError
from flask import current_app
from dataclasses import dataclass
from flask import jsonify
from flask_jwt_extended import jwt_required


@dataclass
class HealthStatus:
    status: str


@dataclass
class HealthError:
    status: str
    error: str


@jwt_required()
def health_check():
    """
    Perform a health check for the application.

    This function checks the health status of the application and returns an appropriate response.
    If the health status is `True`, it returns a JSON response indicating that the application is healthy.
    If the health status is `False`, it raises an `InternalServerError` and returns a JSON response
    indicating that the application is unhealthy.

    Returns:
        Tuple: A tuple containing the JSON response and HTTP status code.

    Raises:
        InternalServerError: Raised when the health status is `False`.

    Example Usage:
    ```
    # Simulate a healthy application
    response, status_code = health_check()
    assert status_code == 200
    assert response == {"status": "healthy"}

    # Simulate an unhealthy application
    response, status_code = health_check()
    assert status_code == 500
    assert response == {"status": "unhealthy", "error": "Health check failed"}
    ```
    """
    try:
        health_status = True  # Change this to False to simulate an error

        if not health_status:
            raise InternalServerError("Health check failed")

        current_app.logger.debug("Health check endpoint was called")

        response = HealthStatus(status="healthy")
        return jsonify(response.__dict__), 200

    except InternalServerError as e:
        current_app.logger.error(f"Health check error: {e}")

        error_response = HealthError(status="unhealthy", error=str(e))
        return jsonify(error_response.__dict__), 500
