requirements.txt
connexion[flask,swagger-ui]==3.0.2
Flask==3.0.0
jsonschema<4.18.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3